% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestProc.R
\name{BT.test_R}
\alias{BT.test_R}
\title{Runs the Barigozzi Trapani test for the number of factors}
\usage{
BT.test_R(X, rmax = 10, alpha = 0.05)
}
\arguments{
\item{X}{(T x N) matrix of observations}

\item{rmax}{maximum number of factors to consider}

\item{alpha}{significance level}
}
\value{
vector with the estimated number of factors
}
\description{
Runs the Barigozzi Trapani test for the number of factors
}
