% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestProc.R
\name{eigenNorm_R}
\alias{eigenNorm_R}
\title{Obtain a standardization factor for the eigenvalues according to Eq. 18}
\usage{
eigenNorm_R(nu, k)
}
\arguments{
\item{nu}{ordered (descending) vector of eigenvalues}

\item{k}{censoring threshold. Largest k-1 eigenvalues are omitted}
}
\value{
Scaling factor
}
\description{
Obtain a standardization factor for the eigenvalues according to Eq. 18
}
