% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestProc.R
\name{randomTest_R}
\alias{randomTest_R}
\title{Runs the randomized testing routine}
\usage{
randomTest_R(phi, p, R, u)
}
\arguments{
\item{phi}{vector of scaled eigenvalues}

\item{p}{pth largest eigenvalue to consider}

\item{R}{Number of simulated random variables}

\item{u}{threshold to define a Bernoulli sequence}
}
\value{
p-value of the test
}
\description{
Runs the randomized testing routine
}
