% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestProc.R
\name{randomTestWrapper_R}
\alias{randomTestWrapper_R}
\title{Runs the iterative randomized testing routine}
\usage{
randomTestWrapper_R(phi, rmax, sig, R = N, u = sqrt(2))
}
\arguments{
\item{phi}{vector of scaled eigenvalues}

\item{rmax}{maximum number of ordered eigenvalues to consider}

\item{sig}{significance level}

\item{R}{Number of simulated random variables per iteration}

\item{u}{threshold to define a Bernoulli sequence}
}
\value{
estimated number of factors
}
\description{
Runs the iterative randomized testing routine
}
