% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{rescaleFactors}
\alias{rescaleFactors}
\title{Calibrates the factors among each other}
\usage{
rescaleFactors(Fmat, Lambda, norm, indx, fd)
}
\arguments{
\item{Fmat}{(T x r_i) matrix holding the factors for i = 1, 2, 3}

\item{Lambda}{(N x r) matrix holding loadings}

\item{norm}{target sum of squares}

\item{indx}{vector indicating the columns of relevant factors}

\item{fd}{logical. If TRUE, the first differences of factors are used}
}
\value{
(T x r_i) matrix with calibrated factors
}
\description{
Calibrates the factors among each other
}
