% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{setsignal2noise}
\alias{setsignal2noise}
\title{Calculate signal to noise ratio}
\usage{
setsignal2noise(U, Lambda, Fmat, drift, F2_indx, F3_indx)
}
\arguments{
\item{U}{(T x N) matrix holding cross-sectional errors}

\item{Lambda}{(N x r) matrix holding loadings}

\item{Fmat}{(T x r) matrix holding the factors}

\item{drift}{logical. If TRUE, a linear trend is included (corresponds to both d_1 and r_1)}

\item{F2_indx}{vector indicating the columns of f2 factors}

\item{F3_indx}{vector indicating the columns of f3 factors}
}
\value{
scaling factor governing the signal to noise ratio
}
\description{
Calculate signal to noise ratio
}
