% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{sim.DGP}
\alias{sim.DGP}
\title{Simulate a DGP according to sec. 5}
\usage{
sim.DGP(
  N = 100,
  Tt = 200,
  drift = TRUE,
  driftI1 = TRUE,
  r_I1 = 2,
  r_I0 = 1,
  factor = FALSE
)
}
\arguments{
\item{N}{number of cross-sectional units}

\item{Tt}{number of simulated time periods}

\item{drift}{logical. If TRUE, a linear trend is included (corresponds to both d_1 and r_1)}

\item{driftI1}{logical. If TRUE, an I(1) factor moves around the linear trend. Else an I(0) factor (corresponds to d_2)}

\item{r_I1}{total number of non zero-mean I(1) factors (corresponds to r_2 + r_1 * d_2)}

\item{r_I0}{number of non zero-mean I(0) factors (corresponds to r_3 + r_1 * (1 - d_2))}

\item{factor}{logical. If TRUE, the factor matrix is returned. Else the simulated observations}
}
\value{
matrix with the simulated observations
}
\description{
Simulate a DGP according to sec. 5
}
