% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simARMACoef}
\alias{simARMACoef}
\title{Simulates the AR and MA coefficients}
\usage{
simARMACoef(pq, limits)
}
\arguments{
\item{pq}{vector holding the lag orders}

\item{limits}{vector holding upper and lower bounds for the coefficients}
}
\value{
vector with stationary simulated coefficients
}
\description{
Simulates the AR and MA coefficients
}
