% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simCrossSecErr}
\alias{simCrossSecErr}
\title{Draws cross-sectional errors according to Eq. 34}
\usage{
simCrossSecErr(Tt, N, a = 0.5, b = 0.5, C = min(floor(N/20), 10))
}
\arguments{
\item{Tt}{number of simulated time periods}

\item{N}{number of cross-sectional units}

\item{a}{parameter governing the serial correlation}

\item{b}{parameter governing the cross-sectional correlation}

\item{C}{parameter governing the extent of the cross-sectional correlation}
}
\value{
(T x N) matrix of cross-sectional errors
}
\description{
Draws cross-sectional errors according to Eq. 34
}
