% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simLambda}
\alias{simLambda}
\title{Draws factor loadings in compliance with Ass. 4}
\usage{
simLambda(r, N)
}
\arguments{
\item{r}{bumber of factors to simulate}

\item{N}{bumber of cross-sectional units}
}
\value{
(N x r) matrix of loadings
}
\description{
Draws factor loadings in compliance with Ass. 4
}
