% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simRW}
\alias{simRW}
\title{Simulates zero-mean I(1) factors according to Eq. 32}
\usage{
simRW(Tt, nBurnin = 1, rholimits = c(0.4, 0.8), sd = 1)
}
\arguments{
\item{Tt}{number of simulated time periods}

\item{nBurnin}{number of burn-in periods for the process}

\item{rholimits}{vector holding upper and lower bounds for the serial correlation parameter}

\item{sd}{of the process innovations}
}
\value{
vector with the simulated factor
}
\description{
Simulates zero-mean I(1) factors according to Eq. 32
}
