% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BTtest}
\alias{BTtest}
\title{Barigozzi & Trapani (\href{https://doi.org/10.1080/07350015.2021.1901719}{2022})  test}
\usage{
BTtest(X, r_max, alpha, BT1)
}
\arguments{
\item{X}{a (T x N) matrix of observations}

\item{r_max}{the maximum number of factors to consider}

\item{alpha}{the significance level}

\item{BT1}{logical. If TRUE, a more conservative eigenvalue rescaling scheme is used}
}
\value{
A vector with the estimated number of (1) factors with a linear trend (2) zero-mean I(1) factors and (3) zero-mean I(0) factors.
}
\description{
Runs the testing routine proposed in Barigozzi & Trapani (\href{https://doi.org/10.1080/07350015.2021.1901719}{2022}) to estimate the number and types of common trends in a nonstationary panel.
}
\details{
For details on the testing procedure I refer to Barigozzi & Trapani (\href{https://doi.org/10.1080/07350015.2021.1901719}{2022}, sec. 4).
}
\examples{
# Simulate a nonstationary panel
# X <- sim_DGP(N = 100, n_Periods = 200)

# Obtain the estimated number of factors (i) with a linear trend (r_1), (ii) zero-mean I(1) (r_2)
# and (iii) zero-mean I(0) (r_3)
# BTtest(X = X, rmax = 10, alpha = 0.05, BT1 = TRUE)
}
\references{
Barigozzi, M., & Trapani, L. (2022). Testing for common trends in nonstationary large datasets. \emph{Journal of Business & Economic Statistics}, 40(3), 1107-1122. DOI: \href{https://doi.org/10.1080/07350015.2021.1901719}{10.1080/07350015.2021.1901719}
}
